+++ src/containerapp/azext_containerapp/tests/latest/utils.py
@@ -28,7 +28,7 @@ def prepare_containerapp_env_for_app_e2e_tests(test_cls, location=TEST_LOCATION)
             if format_location(rg_location) == format_location(STAGE_LOCATION):
                 rg_location = "eastus"
             test_cls.cmd(f'group create -n {rg_name} -l {rg_location}')
-            test_cls.cmd(f'containerapp env create -g {rg_name} -n {env_name} --logs-destination none')
+            test_cls.cmd(f'containerapp env create -g {rg_name} -n {env_name} --logs-destination none', expect_failure=False)
             managed_env = test_cls.cmd('containerapp env show -g {} -n {}'.format(rg_name, env_name)).get_output_in_json()
 
             while managed_env["properties"]["provisioningState"].lower() == "waiting":
@@ -37,15 +37,18 @@ def prepare_containerapp_env_for_app_e2e_tests(test_cls, location=TEST_LOCATION)
     return managed_env["id"]
 
 
-def create_containerapp_env(test_cls, env_name, resource_group, location=None, subnetId=None):
-    logs_workspace_name = test_cls.create_random_name(prefix='containerapp-env', length=24)
-    logs_workspace_location = location
-    if logs_workspace_location is None or format_location(logs_workspace_location) == format_location(STAGE_LOCATION):
-        logs_workspace_location = "eastus"
-    logs_workspace_id = test_cls.cmd('monitor log-analytics workspace create -g {} -n {} -l {}'.format(resource_group, logs_workspace_name, logs_workspace_location), expect_failure=False).get_output_in_json()["customerId"]
-    logs_workspace_key = test_cls.cmd('monitor log-analytics workspace get-shared-keys -g {} -n {}'.format(resource_group, logs_workspace_name)).get_output_in_json()["primarySharedKey"]
+def create_containerapp_env(test_cls, env_name, resource_group, location=None, subnetId=None, needLogsDestination=False):
+    if needLogsDestination:
+        logs_workspace_name = test_cls.create_random_name(prefix='containerapp-env', length=24)
+        logs_workspace_location = location
+        if logs_workspace_location is None or format_location(logs_workspace_location) == format_location(STAGE_LOCATION):
+            logs_workspace_location = "eastus"
 
-    env_command = f'containerapp env create -g {resource_group} -n {env_name} --logs-workspace-id {logs_workspace_id} --logs-workspace-key {logs_workspace_key}'
+        logs_workspace_id = test_cls.cmd('monitor log-analytics workspace create -g {} -n {} -l {}'.format(resource_group, logs_workspace_name, logs_workspace_location), expect_failure=False).get_output_in_json()["customerId"]
+        logs_workspace_key = test_cls.cmd('monitor log-analytics workspace get-shared-keys -g {} -n {}'.format(resource_group, logs_workspace_name)).get_output_in_json()["primarySharedKey"]
+        env_command = f'containerapp env create -g {resource_group} -n {env_name} --logs-workspace-id {logs_workspace_id} --logs-workspace-key {logs_workspace_key}'
+    else:
+        env_command = f'containerapp env create -g {resource_group} -n {env_name} --logs-destination none'
     if location:
         env_command = f'{env_command} -l {location}'